+++ src/main/java/com/ndb/auction/models/BidHolding.java
@@ -6,31 +6,31 @@ import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBDocument;
 @DynamoDBDocument
 public class BidHolding {
     
-    private Long crypto;
-    private Long usd; // crypto * price
+    private Double crypto;
+    private Double usd; // crypto * price
 
     public BidHolding() {
 
     }
 
-    public BidHolding(Long crypto, Long usd) {
+    public BidHolding(Double crypto, Double usd) {
         this.crypto = crypto;
         this.usd = usd;
     }
 
     @DynamoDBAttribute(attributeName = "crypto")
-    public Long getCrypto() {
+    public Double getCrypto() {
         return crypto;
     }
-    public void setCrypto(Long crypto) {
+    public void setCrypto(Double crypto) {
         this.crypto = crypto;
     }
 
     @DynamoDBAttribute(attributeName = "usd")
-    public Long getUsd() {
+    public Double getUsd() {
         return usd;
     }
-    public void setUsd(Long usd) {
+    public void setUsd(Double usd) {
         this.usd = usd;
     }
